requirements.txt
streamlit
streamlit_mic_recorder
openai
langchain
langchain-community
langchain-core
langchain-openai
pydantic==1.10.12
psycopg2-binary
python-dotenv
Werkzeug